Minor bug fixes and more tests.

+++ src/main/backend/indexeddb/JungleDB.js
@@ -68,7 +68,7 @@ class JungleDB {
 
         // Delete existing ObjectStores.
         for (const { tableName, upgradeCondition } of this._objectStoresToDelete) {
-            if (upgradeCondition === null || upgradeCondition === true || upgradeCondition(event.oldVersion, event.newVersion)) {
+            if (upgradeCondition === null || upgradeCondition === true || (typeof upgradeCondition === 'function' && upgradeCondition(event.oldVersion, event.newVersion))) {
                 db.deleteObjectStore(tableName);
             }
         }
@@ -77,7 +77,7 @@ class JungleDB {
         // Create new ObjectStores.
         for (const [tableName, { backend, upgradeCondition }] of this._objectStoreBackends) {
             let IDBobjStore;
-            if (upgradeCondition === null || upgradeCondition === true || upgradeCondition(event.oldVersion, event.newVersion)) {
+            if (upgradeCondition === null || upgradeCondition === true || (typeof upgradeCondition === 'function' && upgradeCondition(event.oldVersion, event.newVersion))) {
                 IDBobjStore = db.createObjectStore(tableName);
             } else {
                 IDBobjStore = request.transaction.objectStore(tableName);
